Move serde_round_trip macro to internals
We currently duplicate the serde_round_trip macro in `units` and
`bitcoin`, this is unnecessary since it is a private test macro we can
just throw it in `internals`.

Note this does mean the macro is publicly available for all builds not
just when testing.

Helps with upcoming crate smashing work also.