Introduce primitives, psbt-v0, bip32, and address
This is not a merge candidate, it is experimental work to try to prove
the state we would like to get to with these three new crates.

By depending on this branch we hope to be able to invert the dependency
relationship between `rust-bitcoin` and `rust-miniscript`.

This is the demo work for epic #2882.