Document installation of cargo-public-api
Installing `cargo-public-api` without using `--locked` leads to a
version that creates spurious changes when checking the public api even
though the version may still be `v0.35.0`, this is surprising to say the
least.

Add documentation to help devs navigate the tool installation. While we
are at it move the `--locked` flag so that the CI command and docs use
the same command incantation.