psbt: Encode keytype as a compact size unsigned integer
Currently we are encoding the `raw::Key`'s type field as a `u8`. This is
not what BIP-174 specifies. Instead it should be a compact size unsigned
integer i.e., a `VarInt`.

Change the field to be a `u64` and encode it using `VarInt`. Introduces
a bunch of casts which are ugly and if someone tries to decode a PSBT
with a keytype larger than a `u8` funny things could happen.

Another solution to this bug would be just to state that we diverge from
the bip and only support key types less than 0xFD - which seems like a
reasonably sane thing to do.

Fix: #2891