bitcoin: Introduce ToU64 conversion trait
Now that we explicitly only support 32 and 64 bit machines we can remove
a bunch of the casts to `u64` by adding a fallible conversion trait.

One of the aims of the trait is to improve the confidence of devs
reading the code, as such it is only for unsigned integers, casting of
signed integers and enums is left as is.