Revamp BIP32 API
The BIP32 API had several problems that we wanted to fix for some time.
It didn't protect the invariants of child numbers, had some
methods fallible even if the errors were cryptographically
impossible, it lacked infallible APIs for deriving xpubs when
absence of hardened children was statically known, and had
annoying ChildNumber enum with named field for no good reason.

In this commit we do a bunch of refactoring to fix these problems.
We add newtypes for the child numbers and a newtype for derivation
path that is guaranteed to only contain normal child numbers with
appropriate parsing and conversion methods. We refactor the
ChildNumber enum to use tuple variants containing the newtypes
and fix the infallible APIs to not return Result. Dealing with
hardened child numbers can still be annoying or non-performant
in some scenarios so we still provide a fallible derivation method.
Further we put the newtypes into separate submodules to make
accidental breaking of invariatns less likely.

Signed-off-by: jaoleal <jgleal@protonmail.com>
Signed-off-by: jaoleal <=>