Remove bech32 from ParseError
In #2381 we attempted to fully encapsulate the `bech32` crate to help
with stabalizing `rust-bitcoin` however we failed to notice the
`address:ParseError` has a variant that includes `bech32`. Public enums
have public variant internals in Rust.

However note that we do not actually return the error ever from the code
because we match on the result of bech32 decoding.

Remove unused error variant from the `address::ParseError` type.

(Note that there are more problems with the parsing because we never
actually return an error when `bech32` decoding fails, moved that
problem to a separate issue: #3044.)

Closes: #3043