Move the CompactTarget type to primitives
Eventually the whole `pow` module will move to `primitives` but this is
not possible easily right now. We would like to be able to move the
`BlockHash` and `block::Header` types over to `primitives` - doing so
requires the `CompactTarget` to be there.

Move the `CompactTarget` type to `primitives` and re-export it from the
`primitives` crate root.

Note also, we re-export the type publicly from `bitcoin::pow`.