Add software impls of `sha` intrinsics for `miri`
The SIMD implementation of SHA256 does some pointer manipulation so it'd
be nice to check it with `miri`. Unfortunately, `miri` currently doesn't
support `sha` intrinsics, so this overrides the missing ones with a
software implemenation. Importnantly, this commit doesn't affect the
pointer manipulation (even though it adds more of it), so testing with
it is still valid.

To actually test this one needs to use these flags:
`RUSTFLAGS="-C target-feature=+sse3,+sse4.1,+sha"`

It should also work on non-x86_64 platforms since `miri` can simulate
x86_64 but you need to use a proper `--target`.