Move script to primitives
Move the script module to the `primitives` crate. This patch is big
and includes a number of things at once but breaking it up is
non-trivial and results in a series of patches that are on their own
unreviewable because they don't make sense without the context of the
other changes or the justification if provided cannot be verified.

Uses an atypical module name `script.rs` instead of `script/mod.rs` so
that the diff is smaller, showing a file move for the move of
`script/mod.rs` to `primitives`. It will be rectified in the next patch.

Notes:

- Adds/updates bitcoinconsensus features as is typical when nested
  down the stack.
- Adds extension traits for things we want to leave behind.
- Makes private methods into private stand alone functions instead of
  adding additional private traits.
- Makes public functions that were defined in the scope of a type (ie
  functions not methods, no `self` parameter) into standalone
  functions.
- Adds a `witness_program` module to group a bunch of these standalone
  functions.
- Leaves the encodable traits behind as expected.
- Makes two `WitnessProgram` constructors public, this increases the
  API surface but is otherwise sane.