impl `Extend` and `FromIterator` on hash types
The hashing-related types lacked standard iterator traits. While it may
not be entirely clear that we consider a hash as a collection of hashed
bytes it makes sense from the trait signatures. An alternative would be
to construct the hash value from bytes but that is impossible because
we would have to construct an array which is fallible. Thus only hashing
makes sense.

This adds `FromIterator` for all non-keyed hash types, `Extend` for all
hash engines and `FromIterator<impl AsRef<[u8; 64]>` for
`sha256::Midstate` since it guarantees length being multiple of 64.