hashes: Move bitcoin-io write impl to macro
For the non-keyed hash types we have a bunch of duplicate calls to
`bitcoin_io::impl_write`. We can reduce duplication by defining an
internal macro in `hashes` that calls the `bitcoin-io` macro.

This adds implementations that are currently missing for `sha512_256`
and `sha384`, as shown by the added unit tests.