hashes: Re-export io crate publicly
We use the `io` crate in the public `sha256t_hash_newtype` macro so it
has to be in scope. Publicly re-export it and use `$crate::io::Foo`.

Also, because of a bug [0] in the `rust-bitcoin-maintainer-tools`
`run_task` script this breakage was not noticed when it was introduced
in https://github.com/rust-bitcoin/rust-bitcoin/pull/3077

[0] https://github.com/rust-bitcoin/rust-bitcoin-maintainer-tools/issues/10