CI: Put fuzz last in CRATES list
There is a bug in `run_task` [0] that drastically reduces our test
coverage. We can bypass it by putting `fuzz` last in the list.

Note, we re-order all crates before `fuzz` to be in alphabetic order,
this is required because when adding `primitives` we poked it in
before `fuzz` to make sure it ran.

[0] rust-bitcoin/rust-bitcoin-maintainer-tools#10