ci: Put fuzz at the end of crates list
There are a couple of problems with the current `run_task` ci script.

We attempted to treat `fuzz` as a special case and only build not test
it. In doing so we (I) wrote `break` instead of `continue` causing the
loop to terminate early and no crates listed after `fuzz` to be tested.

We can ameliorate the problem here by just putting `fuzz` at the end of
the list.

ref: https://github.com/rust-bitcoin/rust-bitcoin-maintainer-tools/issues/10