[io] Add a simple `BufReader`
Since we require `BufRead` for reading most `rust-bitcoin` objects,
its somewhat awkward that we don't provide a `BufReader` in
`bitcoin-io`. This fixes that, providing a simple buffered reader
for use around `bitcoin_io::Read`s.

It differs from `std`'s largely only in missing functionality and
in that it requires a reference to a `Read` rather than a `Read`
directly. This makes it more ergonomic as we don't blanket
implement `Read for &mut Read` unlike `std`, but also allows for
using a `BufReader` without having to take the inner `Read` back
out as its always a reference.

This was built clean-room without looking at `std`/`core2`/etc and
thus has no license requirements.