Add `by_ref` adaptor to IO traits.
In the first released version of `io` we forgot to add a blanket impl
for our IO traits. That significantly degrades usability but even worse,
we cannot add it now because someone might be relying on the absence of
the blanket impl. To resolve the issue we add a wrapper that behaves a
lot like `&mut T`, especially it imlpements the IO traits. Users can now
use it whenever they need to pass a reference to a function accepting
`T: Read` (or other trait). In the next breaking version we will switch
it to real `&mut T` and type alias `ByRef` to make transition smoother.