Provide unbuffered decode API
Since we changed the API to accept `BufRead` there was a push back
against this. As a compromise we implement an alternative API that
allows passing in unbuffered reader without breaking anything. This is
achieved by adding a special method that dispatches the calls as needed,
wrapping the reader in a 1-byte buffer if required.

This design maintains these properties of the `Decodable` trait:
* If no method is overriden it's an infinite recursion
* If any method is overriden everything works, the methods adapt the
  calls as needed
* If multiple methods are overriden (for efficiency) decoding still
  works
* The original methods API stays the same making the change non-breaking

The resulting unbuffered API is slightly annoying to callers which is
acceptable and moderately annoying for implementors which is also
acceptable since there aren't that many imlementors.