Move fmt_to_asm out of extension trait
We currently use the `Script::fmt_to_asm` function to implement `Debug`
on `Script` (calling through to `bytes_to_asm_fmt`).

All of this will be needed over in `primitives` if we want to maintain
the current `Debug` impl.

In preparation for moving `Script` move the `fmt_to_asm` function out of
the extension trait.