Move Script and ScriptBuf to primitives
We are all set up now to move the two script types to `primitives`.

Move `Script` and `ScriptBuf` to the `primitives` crate by doing:

- Add required types to `primitives::prelude`
- Add optional dependency on `hex` to `primitives`
- Move the types
- Re-export them from `blockdata` and the `bitcoin` crate root

Notes on serde

This patch enables `hex` from the `serde` feature. This is a temporary
fix to reduced the size of the diff. A better fix should be worked on
immediately.