Document the new build script magic
This is a follow up to #3182 which introduced a new way of
conditionally including code based on the compiler version.

When originally reviewing I missed the fact that the two loops were
controlled by the current compiler version (`minor`) so the created
macro is different dependent on the compiler used to build the code.

To help the next guy notice, add a comment.