sighash: Make legacy encoding functions take a strong type
We are currently taking a `u32` for the sighash type when we have a
strong type available.

Use `EcdsaSighashType` instead of `u32`.

Note that we just remove the legacy sighash test because all the test
vectors come from libbtc, quoting review:

> The test vectors came from libbtc, which is a minimalist C
  implementation of bitcoin, in 2016. A weird (at best) source of test
  vectors.

The test vectors rely on being able to handle non-standard sighash
types, something we no longer support.

We previously added a regression test to check the legacy sighash logic
with a standard sighash type.