Deprecate Transaction input/output fields
We would like to re-name the `input` and `output` fields of the
`Transaction` struct to be pluralised, however we are trying to be
better library developers and not just break the world all the time.

To help transition, and also to gather feedback on the rename, add
getters using the same name as the proposed new field names. Deprecate
the fields.

Also add `_mut` versions of the getters for ergonomics.