Refactor the key types to hide `secp256k1` crate
We would like to stabilize the key types but the `secp256k1` crate is
unstable and it'll take a long time before it stabilizes. So this change
wraps the types into structs with hopefully more stable API and provides
conversions between them that are backward and forward compatible.

This also happens to make `secp256k1` optional because I thought it'd be
needed although it strictly isn't and I only realized it once it was
almost done, so I decided to keep it in.